Example programs for HW Map + Teleop Holonomic
This commit includes two examples: One HarwareMap, and One TeleOp
Holonomic mode.

The Holonomic Mode is for Mechanum wheels. The robot will move on a
dime. But there are two motor configurations missing. Please fix the
configuration.

The Hardware Map example is where we store and configure the robot
parts. The Holonomic Mode reads the Hardware Map and applies the
movement based off the configurations made there.